requirements.txt
tensorflow-gpu==1.4.1
easydict==1.6
joblib==0.10.3
numpy==1.12.0
opencv-python==3.4.1.15
Pillow==4.0.0
keras==2.0.6
